+++ test/features/sync/gateway/matrix_sdk_gateway_test.dart
@@ -4,6 +4,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:lotti/classes/config.dart';
 import 'package:lotti/features/sync/gateway/matrix_sdk_gateway.dart';
 import 'package:lotti/features/sync/gateway/matrix_sync_gateway.dart';
+import 'package:lotti/features/sync/matrix/sent_event_registry.dart';
 import 'package:matrix/encryption/utils/key_verification.dart';
 import 'package:matrix/matrix.dart';
 import 'package:mocktail/mocktail.dart';
@@ -37,6 +38,7 @@ void main() {
   late StreamController<LoginState> loginStateController;
   late StreamController<KeyVerification> keyVerificationController;
   late bool disposed;
+  late SentEventRegistry sentEventRegistry;
 
   setUp(() {
     client = MockClient();
@@ -54,8 +56,10 @@ void main() {
     when(() => client.dispose()).thenAnswer((_) async {});
 
     disposed = false;
+    sentEventRegistry = SentEventRegistry();
     gateway = MatrixSdkGateway(
       client: client,
+      sentEventRegistry: sentEventRegistry,
       roomStateStream: roomStateController.stream,
       loginStateStream: loginStateController.stream,
       keyVerificationRequestStream: keyVerificationController.stream,